Remove references to sdl and dirs crates from core crate
