+++ nes/Cargo.toml
@@ -6,11 +6,10 @@ edition = "2018"
 
 [dependencies]
 base64 = "0.10"
-dirs = "1.0"
-md-5 = "0.8"
-sdl2 = { version = "0.31", features = ["unsafe_textures"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_bytes = "0.10"
-serde_json = "1.0"
-flate2 = "1.0"
+
+[dev-dependencies]
+md-5 = "0.8"
+sdl2 = { version = "0.31", features = ["unsafe_textures"] }
