+++ nes/src/emulator/state.rs
@@ -1,66 +1,13 @@
 // This file contains the save states API.
 // Changes could break old save states.
 
-use std::fs::{create_dir_all, File};
-use std::path::PathBuf;
-
-use dirs;
-use flate2::read::GzDecoder;
-use flate2::write::GzEncoder;
-use flate2::Compression;
 use serde::{Deserialize, Serialize};
-use serde_json::Serializer;
 
 use crate::emulator::ppu::MirrorMode;
-use crate::emulator::NES;
 
 pub trait SaveState<'de, T: Serialize + Deserialize<'de>> {
     fn freeze(&mut self) -> T;
     fn hydrate(&mut self, t: T);
-}
-
-fn save_state_dir() -> PathBuf {
-    let mut path = match dirs::data_dir() {
-        Some(path) => path,
-        None => panic!("Couldn't get data dir!"),
-    };
-
-    path.push("nes");
-    path.push("save_states");
-    path
-}
-
-fn save_state_file_path(name: &str) -> PathBuf {
-    let mut state_file_path = save_state_dir();
-    state_file_path.push(format!("{}.gz", name));
-    state_file_path
-}
-
-pub fn save_state(nes: &mut NES, name: &str) -> Result<(), String> {
-    create_dir_all(save_state_dir()).map_err(|e| e.to_string())?;
-
-    let state = nes.freeze();
-    let state_file = File::create(save_state_file_path(name)).map_err(|e| e.to_string())?;
-
-    let gzip = GzEncoder::new(state_file, Compression::best());
-    let mut serializer = Serializer::new(gzip);
-    state
-        .serialize(&mut serializer)
-        .map_err(|e| e.to_string())?;
-
-    serializer
-        .into_inner()
-        .try_finish()
-        .map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-pub fn load_state(nes: &mut NES, name: &str) -> Result<(), String> {
-    let state_file = File::open(save_state_file_path(name)).map_err(|e| e.to_string())?;
-    let gzip = GzDecoder::new(state_file);
-    let state = serde_json::from_reader(gzip).map_err(|e| e.to_string())?;
-    nes.hydrate(state);
-    Ok(())
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
