+++ nes/src/emulator/test/image_capture.rs
