+++ nes_sdl/Cargo.toml
@@ -7,4 +7,7 @@ edition = "2018"
 [dependencies]
 nes = { path = "../nes" }
 dirs = "1.0"
+flate2 = "1.0"
+serde = "1.0"
+serde_json = "1.0"
 sdl2 = { version = "0.31", features = ["unsafe_textures"] }
